Fix add_attribute insertion logic
The code didn't work if it couldn't find a property in the file
that is lexicographically greater than the new attribute.
It now accepts _identity as an "end point" to insert before, and
empty lines in _initAttributes() and _useAttributes(), as they mark
the end of the method.

Also add "float" as a possible type, mainly for the "score" attribute
in search results.