Initiate v2
The early history of v2 was squashed in this commit,
mainly because the LGPL license was still everywhere
and I don't want any ambiguity about licensing:
v2 is released under the MIT license.

This also has the positive side effect of removing
large record/replay files for tests that have evolved
a lot before stabilizing.