Don't modify original requestHeaders in Requester.__log()
It's a problem when we need to use the requestHeaders later.

Like in case of __requestRaw() where we send a request + get response,
then __log() it and if the response contains status 301 (redirected)
send another request (with same requestHeaders) to redirected url.

If the __log() erases the 'Authorization' field, the subsequent
request fails.

Fixes #470