No longer blindly mask JSON parsing errors
If the JSON returned fails to parse *for whatever reason*, it is wrapped
into a dictionary and blindly passed on. Where, for example, PaginatedList
then returns a mostly unpopulated object and raises an URL exception
when attempting to complete it. This has wasted my time for the last
time. Clean up all of the unparseble JSON in the replay data (which is
only fair, since I added most of it), and shift the exception handling
up one level.