Re-build tests/ExportAllAttributes.py
The change to last_modified for ContentFile has introduced an
additional API request. This breaks the recorded requests in
the replay data for this test. The only way to fix this is to
re-build the testcase and its replaydata. Now the test itself
is failing because a lot of Classes are out of sync with the
Github API.