Use generic for expected_type in BadAttributeException
The expected_type return value is bound to expected_type argument on
init.  This will type the return of expected_type to the type passed to
the init function.

I'm choosing to not bind TypeVar to anything explicit, as in theory
the exception should not be limited to any specific expected type, and
the union quickly balloons and was already significant.

This fixes the implicit Any generic dict in Type[dict], that turns into
Type[dict[Any, Any]] and is disallowed when setting
--disallow-any-generics (this still allows explicitly defined Any
generics).  The generic also means consumption of the returned value
doesn't require handling all listed types in the Union as the type will
be specifically the type of expected_type.

Included is a couple lines of black formatting as they occurred on
write and are insignificant in number.