Fix handling of `create_git_blob`'s result
`create_git_blob` does not actually return the entire blob object
content, per the OpenAPI description it only returns a `short-blob`
with the `sha` and `url` properties whereas a "full" blob also has
`encoding`, `content`, and `size` (and `node_id` but that's not
relevant).

Fix by marking the `GitBlob` returned as non-complete, this way normal
completion will fetch the additional data iff it is accessed.