Introduce `assertWarningIn` helper
In some cases, some deprecation messages may be generated by third-party libraries depedending
on the python version we are using.

This is a problem when we want to insure that a specific message is raised.

This change introduce `assertWarningIn`, which will check if a specific deprecation warning
is in the list of warnings.