Port GitHub timestamp parsing away from python-dateutil
According to https://docs.github.com/en/rest/overview/resources-in-the-rest-api#schema
the GitHub API only returns timestamps in the format YYYY-MM-DDTHH:MM:SSZ which
we can parse with the Python stdlib back to 3.7 if we handle the "Z" ourselves.

This removes the test cases using fractional seconds since the stdlib (< 3.11) can't
parse them and according to the above GitHub doesn't return them anyway.