Port last_modified_datetime parsing away from python-dateutil
Dates in HTTP are currently specified in RFC 7231, which references
RFC 5322 for the format, which is an updated version of RFC 2822.
To replace dateutil we can use the email.utils.parsedate_to_datetime()
stdlib function, which parses RFC 2822 dates.

It also parses the obsolete variants without a timezone and returns
timezone naive datetime objects. RFC 7231 states that UTC should be
assumed in such a case, so convert it to a timezone aware instance
with UTC in such a case.

This is the same way werkzeug parses HTTP dates, see
https://werkzeug.palletsprojects.com/en/3.0.x/http/#werkzeug.http.http_date