original hack did not work, try a worse one
Setting TYPE_CHECKING apparently triggers a circular import error in urllib3 which is not great.

Try an alternate option, checking if sphinx is imported. This is just to investigate if circular imports are an issue, as this would not be acceptable to release (there might be people using pygithub within sphinx for instance)