Repo's Advanced Security Flag Is Not Deprecated
It turns out that these advanced security features are sometimes not
returned due to your licensing config (if you disable GHAS then the
'advanced_security' bit of a repository's 'security_and_analysis' is
just not included). There may be more features here as the documentation
doesn't include some of these but I have personally seen them (in this
case, that would just be 'secret_scanning_validity_checks' which appears
to be a limited rollout at the moment but might as well support it).