Add `has_discussions` to `AuthenticatedUser` and `Repository` classes (#3020)
The `has_discussions` feature is exposed in the Github API and is
identical in the way it works to something like `has_wiki`. Following a
couple previous PR attempts by @amnuts and @heitorPB as well as the
existing code for `has_wiki` I added `has_discussions` as well as the
test cases.

Fixes #2733 and #2994