Use warnings.deprecated() decorator with Python 3.13+
PyGitHub is using the "deprecated" Python package for deprecations
until now, but was only using the provided deprecated() decorator
from it, which in Python 3.13 has a stdlib equivalent with
warnings.deprecated().

The stdlib variant has multiple advantages:

* Removes the dependency on the "deprecated" package and the
  transitive "wrapt" package (which is a C extension)
* Mypy can alert users of deprecated APIs since mypy 1.14 (opt-in),
  see https://mypy-lang.blogspot.com/2024/12/mypy-114-released.html
* Pylance in VSCode gives visual hints and warnings when using
  a deprecated function/class

The make everything still work with older Python we provide a wrapper
that implements the new stdlib interface, and we only require the
"deprecated" package with older Python. The wrapper requires the
"extra_stacklevel" kwarg added in 1.2.15, so make that the minimum version.
The wrapper exposed a bug in types-deprecated, which is tracked
here: https://github.com/python/typeshed/pull/13573 which is why
type checking is skipped in parts there.

For classes the stdlib variants just warns with the provided message
while the package converts it to "Call to deprecated class Foo (<msg>)".
Adjust the tests to use regex matching there, to cover both cases.