Throw the right exception for missing branch
Fixes #3011

Ideally this code should just check for status code and not touch the
message at all, but I'm sticking with the pattern here.

Updated the tests to reflect the new exception:

1. Replaced base `github.GithubException` with the more specific `github.UnknownObjectException`

2. Removed message check `self.assertEqual(raisedexp.exception.message, "Branch not found")` in tests

createException() doesn't pass a message to any specific exceptions, so
this change is in line with other errors like `BadCredentialsException`
or `RateLimitExceededException` which also don't have a message field
set.