Add `delete_self_hosted_runner` to `Organization` (#3306)
PR related to #3189

This PR introduces a new function `delete_self_hosted_runner` to handle
the deletion of self-hosted runners at the organization level.
Additionally, it includes a documentation update to correct the
description of the `get_self_hosted_runners` function.

For guidance on tests, I referred to PRs #3190 and #3203 (huge thanks to
@climbfuji and @billnapier)

---------

Signed-off-by: Pavel Abramov <uncle.decart@gmail.com>
Co-authored-by: Enrico Minack <github@enrico.minack.dev>