Point deprecation warnings to the caller code rather than inner class (#3275)
By default `warnings.warn` will point at the location of the `warn` call
itself, and provides no stacktrace information, which can make it much
harder to find the relevant application code.

By setting the stacklevel to 2, `warnings` will provide a link (file and
line number) to the *caller* of `Github()`, which is the location to
fix.